refactor: BowlingGame 두개의 인터페이스를 구현하도록 변경
