+++ src/main/java/bowling/domain/bowlinggame/BowlingGameService.java
@@ -1,0 +1,11 @@
+package bowling.domain.bowlinggame;
+
+import bowling.domain.KnockDownPins;
+
+public interface BowlingGameService{
+    void setKnockDownPins(KnockDownPins knockDownPins);
+
+    boolean isEnd();
+
+    int getCurrentFrameIndex();
+}
