+++ src/main/java/bowling/domain/bowlinggame/BowlingGameViewDto.java
@@ -1,0 +1,12 @@
+package bowling.domain.bowlinggame;
+
+import bowling.domain.PlayerName;
+import bowling.domain.frame.Frame;
+
+import java.util.stream.Stream;
+
+public interface BowlingGameViewDto {
+    Stream<Frame> framesStream();
+
+    PlayerName getPlayerName();
+}
