+++ src/main/java/bowling/view/ResultView.java
@@ -1,7 +1,7 @@
 package bowling.view;
 
-import bowling.domain.BowlingGame;
+import bowling.domain.bowlinggame.BowlingGameViewDto;
 
 public interface ResultView {
-    void print(BowlingGame bowlingGame);
+    void print(BowlingGameViewDto bowlingGame);
 }
